conn

+++ internal/domain/conn/imp/user.go
@@ -7,12 +7,12 @@ import (
 )
 
 func NewUser(userName string, conn jinx_int.IConnection, userManager user_int.IUserManager) user_int.IUser {
-	userId := rand.Uint32()
-	return &user{userId: userId, userName: userName, conn: conn, userManager: userManager}
+	userId := rand.Uint64()
+	return &user{userId: uint(userId), userName: userName, conn: conn, userManager: userManager}
 }
 
 type user struct {
-	userId      uint32
+	userId      uint
 	userName    string
 	conn        jinx_int.IConnection
 	userManager user_int.IUserManager
@@ -31,7 +31,7 @@ func (u *user) GetUserIP() string {
 }
 
 func (u *user) Offline() {
-	u.userManager.RemoveUser(u.userId)
+	u.userManager.RemoveUser(uint32(u.userId))
 }
 
 func (u *user) Online() {
